+++ ghost/constants/index.js
@@ -1,18 +1,1 @@
-module.exports = {
-    ONE_HOUR_S: 3600,
-    ONE_DAY_S: 86400,
-    ONE_WEEK_S: 604800,
-    ONE_MONTH_S: 2628000,
-    SIX_MONTH_S: 15768000,
-    ONE_YEAR_S: 31536000,
-    FIVE_MINUTES_MS: 300000,
-    ONE_HOUR_MS: 3600000,
-    ONE_DAY_MS: 86400000,
-    ONE_WEEK_MS: 604800000,
-    ONE_MONTH_MS: 2628000000,
-    SIX_MONTH_MS: 15768000000,
-    ONE_YEAR_MS: 31536000000,
-    STATIC_IMAGES_URL_PREFIX: 'content/images',
-    STATIC_MEDIA_URL_PREFIX: 'content/media',
-    STATIC_FILES_URL_PREFIX: 'content/files'
-};
+module.exports = require('./lib/constants');
