+++ ghost/custom-theme-settings-service/package.json
@@ -9,7 +9,9 @@
     "dev": "echo \"Implement me!\"",
     "test:unit": "NODE_ENV=testing c8 --all --reporter text --reporter cobertura --check-coverage mocha './test/**/*.test.js'",
     "test": "yarn test:unit",
-    "lint": "eslint . --ext .js --cache"
+    "lint:code": "eslint *.js lib/ --ext .js --cache",
+    "lint": "yarn lint:code && yarn lint:test",
+    "lint:test": "eslint -c test/.eslintrc.js test/ --ext .js --cache"
   },
   "files": [
     "index.js",
