+++ ghost/oembed-service/test/hello.test.js
@@ -2,6 +2,6 @@ const assert = require('assert');
 
 describe('Hello world', function () {
     it('Runs a test', function () {
-        assert.equal('hello', 'hello');
+        assert.ok(require('../index'));
     });
 });
